p2p: rename network folder to qvic
